10th level - Clockwork
Added a 10th level, completing my goal for the minimum for this pack. I might add more, who knows? (I probably will though because I get bored a lot and like doing this sort of thing). That being said, this (potentially) final level is quite a challenge, and a bit more unique than the other levels.